+++ yolox/object_detection_config.yaml
@@ -6,12 +6,12 @@ val_file: train.json
 
 TRAIN_SIZE: 32
 VAL_SIZE: 8
-LOSS_IOU_TH: 0.3
+UNLABELED_SIZE: 8
 
+IOU_TH: 0.3
 BACKGROUND_LABEL: 81
 CLASSES: 80
 CONF_THRESH: 0.35
-CM_IOU_THRESH: 0.25
 TOP_K: 100
 IMAGE_SIZE: [640, 640]
 SCALE_TO_SIZE: 636
